Rewrite gameserverstatus to support slash and components v2
