+++ gameserverstatus/utils/gameservers.py
@@ -1,0 +1,100 @@
+QSTAT_TYPES = {
+    "a2s": "Half-Life 2 new server",
+    "alienarenas": "Alien Arena server",
+    "ams": "America's Army v2.x server",
+    "bfbc2": "Battlefield Bad Company 2 server",
+    "bfs": "BFRIS server",
+    "cod2s": "Call of Duty 2 server",
+    "cod4s": "Call of Duty 4 server",
+    "cods": "Call of Duty server",
+    "crs": "Command and Conquer: Renegade server",
+    "crysis": "Crysis server",
+    "cubes": "Sauerbraten server",
+    "d3g": "Descent3 Gamespy Protocol server",
+    "d3p": "Descent3 PXO protocol server",
+    "d3s": "Descent3 server",
+    "dirtybomb": "DirtyBomb server",
+    "dm3s": "Doom 3 server",
+    "efm": "Star Trek: Elite Force server",
+    "efs": "Star Trek: Elite Force server",
+    "etqws": "QuakeWars server",
+    "eye": "All Seeing Eye Protocol server",
+    "farmsim": "FarmingSimulator server",
+    "fcs": "FarCry server",
+    "fls": "Frontlines-Fuel of War server",
+    "gps": "Gamespy Protocol server",
+    "grs": "Ghost Recon server",
+    "gs2": "Gamespy V2 Protocol server",
+    "gs3": "Gamespy V3 Protocol server",
+    "gs4": "Gamespy V4 Protocol server",
+    "h2s": "Hexen II server",
+    "hazes": "Haze Protocol server",
+    "hl2s": "Half-Life 2 server",
+    "hla2s": "Half-Life server",
+    "hlqs": "Half-Life server",
+    "hls": "Half-Life server",
+    "hrs": "Heretic II server",
+    "hws": "HexenWorld server",
+    "iourts": "Urban Terror server",
+    "jk2m": "Jedi Knight 2 server",
+    "jk2s": "Jedi Knight 2 server",
+    "jk3m": "Jedi Knight: Jedi Academy server",
+    "jk3s": "Jedi Knight: Jedi Academy server",
+    "kps": "Kingpin server",
+    "ksp": "Kerbal Space Program server",
+    "maqs": "Medal of Honor: Allied Assault (Q) server",
+    "mas": "Medal of Honor: Allied Assault server",
+    "mhs": "Medal of Honor: Allied Assault server",
+    "mumble": "Mumble server",
+    "netp": "NetPanzer server",
+    "nexuizs": "Nexuiz server",
+    "openarenas": "OpenArena server",
+    "ottds": "OpenTTD server",
+    "preys": "PREY server",
+    "prs": "Pariah server",
+    "q2s": "Quake II server",
+    "q3rallys": "Q3 Rally server",
+    "q3s": "Quake III: Arena server",
+    "q4s": "Quake 4 server",
+    "qs": "Quake server",
+    "quetoos": "Quetoo server",
+    "qws": "QuakeWorld server",
+    "reactions": "Reaction server",
+    "rss": "Ravenshield server",
+    "rws": "Return to Castle Wolfenstein server",
+    "sas": "Savage server",
+    "sfs": "Soldier of Fortune server",
+    "sgs": "Shogo: Mobile Armor Division server",
+    "smokingunss": "Smokin' Guns server",
+    "sms": "Serious Sam server",
+    "sns": "Sin server",
+    "sof2s": "Soldier of Fortune 2 server",
+    "starmade": "StarMade server",
+    "t2s": "Tribes 2 server",
+    "tbs": "Tribes server",
+    "tees": "Teeworlds server",
+    "terraria": "Terraria server",
+    "tf": "Titanfall server",
+    "tf2": "Titanfall 2 server",
+    "tf2e": "Titanfall 2 Protocol v2 server",
+    "tm": "TrackMania server",
+    "tremulousgpps": "Tremulous GPP server",
+    "tremulouss": "Tremulous server",
+    "ts2": "Teamspeak 2 server",
+    "ts3": "Teamspeak 3 server",
+    "turtlearenas": "Turtle Arena server",
+    "uns": "Unreal server",
+    "unvanquisheds": "Unvanquished server",
+    "ut2004s": "UT2004 server",
+    "ut2s": "Unreal Tournament 2003 server",
+    "ut3s": "UT3 server",
+    "vent": "Ventrilo server",
+    "warsows": "Warsow server",
+    "waws": "Call of Duty World at War server",
+    "wics": "World in Conflict server",
+    "woets": "Enemy Territory server",
+    "wolfs": "Wolfenstein server",
+    "wops": "World Of Padman server",
+    "xonotics": "Xonotic server",
+    "zeq2lites": "ZEQ2 Lite server",
+}
